requirements.txt
# ==========================================
# Core Framework
# ==========================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ==========================================
# LLM & AI
# ==========================================
openai==1.3.0
anthropic==0.7.1
langchain==0.0.340
langchain-openai==0.0.2
langchain-anthropic==0.0.1
langgraph==0.0.20
dspy-ai==2.4.0

# ==========================================
# Topic Modeling (BERTopic)
# ==========================================
bertopic==0.16.0
sentence-transformers==2.2.2
umap-learn==0.5.5
hdbscan==0.8.33

# ==========================================
# Vector Database (Weaviate)
# ==========================================
weaviate-client==3.25.3

# ==========================================
# Graph Database (Neo4j)
# ==========================================
neo4j==5.14.1

# ==========================================
# Event Streaming (Kafka)
# ==========================================
faust-streaming==0.10.0
kafka-python==2.0.2
aiokafka==0.8.1

# ==========================================
# Databases & Caching
# ==========================================
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.1
hiredis==2.2.3
sqlalchemy==2.0.23

# ==========================================
# Observability (LangFuse)
# ==========================================
langfuse==2.6.3

# ==========================================
# Integrations
# ==========================================
# ServiceNow
pysnow==0.7.17
# Jira
jira==3.5.2
# GitHub
PyGithub==2.1.1
# Slack (for HITL)
slack-sdk==3.23.0

# ==========================================
# Infrastructure as Code
# ==========================================
python-terraform==0.10.1
ansible-runner==2.3.4

# ==========================================
# Data Processing
# ==========================================
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2

# ==========================================
# Utilities
# ==========================================
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1
python-multipart==0.0.6

# ==========================================
# Testing
# ==========================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# ==========================================
# Cloud Providers (GCP)
# ==========================================
google-cloud-monitoring==2.16.0
google-cloud-compute==1.14.1
google-cloud-logging==3.8.0
google-auth==2.25.2

# ==========================================
# Monitoring & Logging
# ==========================================
prometheus-client==0.19.0
structlog==23.2.0
